Update test car detection.py
Jay: A few bug fixes to in accordance to the change in how the AI outputs. 
(Original model was not using dense layers therefore creating a "Too many indicies" error.)

fixed the input menu where the responce would end up at the end of "Video: 3" instead of on a new line due to a typo.